first steps of correcting tests
