+++ mega-zep-backend/src/main/java/com/gepardec/mega/service/impl/monthlyreport/calculation/journey/JourneyDirectionHandler.java
@@ -10,26 +10,29 @@ public class JourneyDirectionHandler {
     private JourneyDirection beforeJourneyDirection = JourneyDirection.BACK;
 
     public Optional<Warning> moveTo(JourneyDirection journeyDirection) {
+        Optional<Warning> warning = Optional.empty();
         switch (journeyDirection) {
             case TO_AIM:
                 if (beforeJourneyDirection == JourneyDirection.TO_AIM || beforeJourneyDirection == JourneyDirection.FURTHER) {
-                    return Optional.of(Warning.WARNING_JOURNEY_BACK_MISSING);
+                    warning = Optional.of(Warning.WARNING_JOURNEY_BACK_MISSING);
                 }
                 break;
             case FURTHER:
             case BACK:
                 if (beforeJourneyDirection == JourneyDirection.BACK) {
-                    return Optional.of(Warning.WARNING_JOURNEY_TO_AIM_MISSING);
+                    beforeJourneyDirection = journeyDirection;
+                    warning = Optional.of(Warning.WARNING_JOURNEY_TO_AIM_MISSING);
                 }
                 break;
             case INVALIDATE:
                 resetHandler();
-                return Optional.of(Warning.WARNING_JOURNEY_BACK_MISSING);
+                warning = Optional.of(Warning.WARNING_JOURNEY_BACK_MISSING);
+                break;
             default:
                 throw new IllegalArgumentException("Enum type '" + journeyDirection.name() + "' not supported!");
         }
         beforeJourneyDirection = journeyDirection;
-        return Optional.empty();
+        return warning;
     }
 
     public boolean isJourneyFinished() {
